Reference

Contents

Reference

Contents

Index

Index

ProxTV.NormLp

ProxTV.NormTVp

ProxTV.ShiftedNormLp

ProxTV.ShiftedNormTVp

ProxTV.TVp_norm

ProximalCore.prox!

ProximalCore.prox!

ProximalCore.prox!

ProximalCore.prox!

ProximalCore.prox!

ShiftedProximalOperators.shift!

ShiftedProximalOperators.shift!

ShiftedProximalOperators.shifted

ShiftedProximalOperators.shifted

ShiftedProximalOperators.shifted

ShiftedProximalOperators.shifted

ShiftedProximalOperators.shifted

ProxTV.NormLp — Type

NormLp(λ::Real or AbstractArray, p::Real)

Represents the Lp norm with parameter p and scaling factor λ.

ProxTV.NormTVp — Type

NormTVp(λ::Real or AbstractArray, p::Real)

Represents the Total Variation (TV) norm with parameter p and scaling factor λ.

ProxTV.ShiftedNormLp — Type

ShiftedNormLp

A mutable struct representing a shifted NormLp function.

ProxTV.ShiftedNormTVp — Type

ShiftedNormTVp

A mutable struct representing a shifted NormTVp function.

ProxTV.TVp_norm — Method

TVp_norm(x::AbstractArray, p::Real)

Computes the TVp norm of vector x with parameter p.

ProximalCore.prox! — Method

prox!(y, h::NormLp, q, ν, ctx_ptr, callback)

Evaluates inexactly the proximity operator of a Lp norm object. The duality gap at the solution is guaranteed to be less than dualGap.

Inputs: - y: Array in which to store the result. - h: NormLp object. - q: Vector to which the proximity operator is applied. - ν: Scaling factor. - ctx_ptr: Pointer to the context object. - callback: Pointer to the callback function.

ProximalCore.prox! — Method

prox!(y, h::NormTVp, q, ν; dualGap=1e-5)

Evaluates inexactly the proximity operator of a TVp norm object. The duality gap at the solution is guaranteed to be less than dualGap.

Inputs: - y: Array in which to store the result. - h: NormTVp object. - q: Vector to which the proximity operator is applied. - ν: Scaling factor. - ctx_ptr: Pointer to the context object. - callback: Pointer to the callback function.

ProximalCore.prox! — Method

prox!(y, ψ::ShiftedNormLp, q, ν, ctx_ptr, callback)

Evaluates inexactly the proximity operator of a shifted Lp norm. The duality gap at the solution is guaranteed to be less than dualGap.

Inputs: - y: Array in which to store the result. - ψ: ShiftedNormLp object. - q: Vector to which the proximity operator is applied. - ν: Scaling factor. - ctx_ptr: Pointer to the context object. - callback: Pointer to the callback function.

ProximalCore.prox! — Method

prox!(y, ψ::ShiftedNormTVp, q, σ, ctx_ptr, callback)

Evaluates inexactly the proximity operator of a shifted TVp norm. The duality gap at the solution is guaranteed to be less than dualGap.

Inputs: - y: Array in which to store the result. - ψ: ShiftedNormTVp object. - q: Vector to which the proximity operator is applied. - ν: Scaling factor. - ctx_ptr: Pointer to the context object. - callback: Pointer to the callback function.

ProximalCore.prox! — Method

prox!(y, ψ::Union{InexactShiftedProximableFunction, ShiftedProximableFunction}, q, ν; ctx_ptr, callback)

Evaluates the proximity operator of a shifted regularizer, choosing between exact and inexact calculations based on the type of ψ and the presence of pointers.

If ψ is a ShiftedProximableFunction and pointers are not provided, computes the exact proximity operator.

If ψ is an InexactShiftedProximableFunction and pointers are provided, computes the inexact proximity operator by calling the callback.

Inputs: - y: Array in which to store the result. - ψ: Either a ShiftedProximableFunction (for exact prox) or an InexactShiftedProximableFunction (for inexact prox). - q: Vector to which the proximity operator is applied. - ν: Scaling factor. - ctx_ptr: Pointer to the context object. - callback: Pointer to the callback function. Outputs: - The solution is stored in the input vector y, which is also returned.

Errors: - Raises an error if ψ is of type ShiftedProximableFunction and pointers are provided, or if ψ is of type InexactShiftedProximableFunction and pointers are not provided.

ShiftedProximalOperators.shift! — Method

shift!(ψ::ShiftedNormLp, shift::AbstractVector)

Updates the shift of a ShiftedNormLp object.

ShiftedProximalOperators.shift! — Method

shift!(ψ::ShiftedNormTVp, shift::AbstractVector)

Updates the shift of a ShiftedNormTVp object.

ShiftedProximalOperators.shifted — Method

shifted(h::Union{NormLp, NormTVp}, xk::AbstractVector)

Creates a shifted version of h depending on its type. If h is of type NormLp, returns a ShiftedNormLp. If h is of type NormTVp, returns a ShiftedNormTVp.

ShiftedProximalOperators.shifted — Method

shifted(h::NormLp, xk::AbstractVector)

Creates a ShiftedNormLp object with initial shift xk.

ShiftedProximalOperators.shifted — Method

shifted(h::NormTVp, xk::AbstractVector)

Creates a ShiftedNormTVp object with initial shift xk.

ShiftedProximalOperators.shifted — Method

shifted(ψ::ShiftedNormLp, sj::AbstractVector)

Creates a ShiftedNormLp object by adding a second shift sj.

ShiftedProximalOperators.shifted — Method

shifted(ψ::ShiftedNormTVp, sj::AbstractVector)

Creates a ShiftedNormTVp object by adding a second shift sj.